Added expected hash support for data sealing
Added the ability to seal data to expected values of PCRs instead of
current values using the -f/--permfile argument.  The format of the
argument is similar to that seen in the tpm_nvdefine command, with the
exception of the r/w leading character (with sealing of data, it's
implicity 'r').

This can be helpful if you need to seal data to be read at next boot
time when you know the PCRs will be in a different state.  This can
happen when updating the kernel and/or bootloader options using a
TPM-aware bootloader.

If a PCR is given using the -p option and again in the -f, the -f should
override.